requirements.txt
# Core dependencies
openai>=1.50.0
pyyaml>=6.0
tenacity>=8.2.0

# Database
psycopg2-binary>=2.9.0

# Storage
boto3>=1.28.0

# Monitoring
psutil>=5.9.0

# Optional: Document parsing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Environment
python-dotenv>=1.0.0

# Web Server
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
